refactor: rename directory for accordions
