+++ components/accordions/Pattern.tsx
