+++ components/dialogs/ClothingDetailsModal.tsx
@@ -11,13 +11,13 @@ import {
 } from "react-native";
 import CloseModalIcon from "../../assets/icons/modal/close-modal.svg";
 import DesignIcon from "../../assets/icons/tabs/design.svg";
-import SeasonAccordion from "../clothing-detail-accordion/Season";
-import OccasionSelection from "../clothing-detail-accordion/Occasion";
-import CategorySelection from "../clothing-detail-accordion/Category";
-import ColorAccordion from "../clothing-detail-accordion/Color";
+import SeasonAccordion from "../accordions/Season";
+import OccasionSelection from "../accordions/Occasion";
+import CategorySelection from "../accordions/Category";
+import ColorAccordion from "../accordions/Color";
 import Spinner from "../common/Spinner";
-import MaterialAccordion from "../clothing-detail-accordion/Material";
-import PatternAccordion from "../clothing-detail-accordion/Pattern";
+import MaterialAccordion from "../accordions/Material";
+import PatternAccordion from "../accordions/Pattern";
 import { updateClothing } from "@/network/web/clothes";
 import Toast from "react-native-toast-message";
 import { useUser } from "../config/user-context";
